+++ packages/design-system-web/.build/stencil.bindings.react.ts
@@ -1,0 +1,13 @@
+import { reactOutputTarget as react } from "@stencil/react-output-target";
+
+
+export const reactGenerator = (
+    componentCorePackage = '@geovistory/design-system-web',
+    loaderDir = '/loader',
+    proxiesFile = '../design-system-react/src/components/stencil-generated/index.ts',
+) => react({
+        componentCorePackage: componentCorePackage,
+        proxiesFile: proxiesFile,
+        includeDefineCustomElements: true,
+        loaderDir
+    })