+++ packages/design-system-web/.storybook/main.js
@@ -1,13 +1,16 @@
 module.exports = {
   "stories": [
     "../docs/**/*.stories.mdx",
-    "../src/**/*.stories.mdx",
+    "../docs/**/*.stories.@(js|jsx|ts|tsx)",
     "../src/**/*.stories.@(js|jsx|ts|tsx)"
   ],
   "addons": [
+    'storybook-design-token',
     "@storybook/addon-links",
     "@storybook/addon-essentials",
-    'storybook-design-token'
+    "@storybook/addon-interactions"
   ],
-  "framework": "@storybook/html"
+  "framework": "@storybook/react",
+  "staticDirs": ["../dist/design-system-web"],
+
 }