+++ packages/design-system-web/src/components/advanced/geov-navbar/geov-navbar.stories.tsx
@@ -1,8 +1,0 @@
-export default {
-    title: 'Advanced Components/Navbar',
-}
-
-const Template = () => `
-`
-
-export const Navbar = Template.bind({});