+++ packages/design-system-web/CHANGELOG.md
@@ -3,6 +3,142 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 1.1.0 (2022-07-13)
+
+**Note:** Version bump only for package @geovistory/design-system-web
+
+
+
+
+
+## 1.0.5-pr-16.1 (2022-07-13)
+
+**Note:** Version bump only for package @geovistory/design-system-web
+
+
+
+
+
+## 1.0.5-pr-16.0 (2022-07-13)
+
+**Note:** Version bump only for package @geovistory/design-system-web
+
+
+
+
+
+## 1.0.4 (2022-07-08)
+
+**Note:** Version bump only for package @geovistory/design-system-web
+
+
+
+
+
+## 1.0.3 (2022-07-07)
+
+**Note:** Version bump only for package @geovistory/design-system-web
+
+
+
+
+
+## 1.0.2 (2022-07-07)
+
+**Note:** Version bump only for package @geovistory/design-system-web
+
+
+
+
+
+## [1.0.1](https://github.com/geovistory/design-system/compare/v1.0.0...v1.0.1) (2022-07-07)
+
+**Note:** Version bump only for package @geovistory/design-system-web
+
+
+
+
+
+## 1.0.0 (2022-07-07)
+
+**Note:** Version bump only for package @geovistory/design-system-web
+
+
+
+
+
+## 0.3.1-pr-14.8 (2022-07-07)
+
+**Note:** Version bump only for package @geovistory/design-system-web
+
+
+
+
+
+## 0.3.1-pr-14.7 (2022-07-07)
+
+**Note:** Version bump only for package @geovistory/design-system-web
+
+
+
+
+
+## 0.3.1-pr-14.6 (2022-07-07)
+
+**Note:** Version bump only for package @geovistory/design-system-web
+
+
+
+
+
+## 0.3.1-pr-14.5 (2022-07-06)
+
+**Note:** Version bump only for package @geovistory/design-system-web
+
+
+
+
+
+## 0.3.1-pr-14.4 (2022-07-06)
+
+**Note:** Version bump only for package @geovistory/design-system-web
+
+
+
+
+
+## 0.3.1-pr-14.3 (2022-07-06)
+
+**Note:** Version bump only for package @geovistory/design-system-web
+
+
+
+
+
+## 0.3.1-pr-14.2 (2022-07-05)
+
+**Note:** Version bump only for package @geovistory/design-system-web
+
+
+
+
+
+## 0.3.1-pr-14.1 (2022-07-01)
+
+**Note:** Version bump only for package @geovistory/design-system-web
+
+
+
+
+
+## 0.3.1-pr-14.0 (2022-06-28)
+
+**Note:** Version bump only for package @geovistory/design-system-web
+
+
+
+
+
 ## 0.3.0 (2022-06-17)
 
 **Note:** Version bump only for package @geovistory/design-system-web